refactor(gallery): split default db service
split default db service to users and images

MODULE3-PART2-dynamodb
